mv spirv module to its own file
