+++ src/runtime/opencl/opencl_module.cc
@@ -30,7 +30,6 @@
 #include <vector>
 
 #include "../source_utils.h"
-#include "../spirv/spirv_shader.h"
 #include "opencl_common.h"
 
 namespace tvm {
@@ -107,26 +106,6 @@ class OpenCLWrappedFunc {
   std::vector<size_t> arg_size_;
   // launch parameters config
   LaunchParamConfig launch_param_config_;
-};
-
-class OpenCLSPIRVModuleNode : public OpenCLModuleNodeBase {
- public:
-  explicit OpenCLSPIRVModuleNode(const std::unordered_map<std::string, SPIRVShader>& shaders,
-                                 const std::string& spirv_text,
-                                 std::unordered_map<std::string, FunctionInfo> fmap)
-      : OpenCLModuleNodeBase(fmap), shaders_(shaders), spirv_text_(spirv_text) {}
-
-  void SaveToFile(const std::string& file_name, const std::string& format) final;
-  void SaveToBinary(dmlc::Stream* stream) final;
-  std::string GetSource(const std::string&) final { return spirv_text_; }
-
-  void Init() override;
-  cl_kernel InstallKernel(cl::OpenCLWorkspace* w, cl::OpenCLThreadEntry* t,
-                          const std::string& func_name, const KTRefEntry& e) override;
-
- private:
-  std::unordered_map<std::string, SPIRVShader> shaders_;
-  std::string spirv_text_;
 };
 
 OpenCLModuleNodeBase::~OpenCLModuleNodeBase() {
@@ -250,7 +229,7 @@ cl_kernel OpenCLModuleNode::InstallKernel(cl::OpenCLWorkspace* w, cl::OpenCLThre
       programs_[func_name][device_id] =
           clCreateProgramWithSource(w->contexts[platform], 1, &s, &len, &err);
       OPENCL_CHECK_ERROR(err);
-    } else if (fmt_ == "xclbin" || fmt_ == "awsxclbin" || fmt_ == "aocx" || fmt_ == "spirv") {
+    } else if (fmt_ == "xclbin" || fmt_ == "awsxclbin" || fmt_ == "aocx") {
       const unsigned char* s = (const unsigned char*)data_.c_str();
       size_t len = data_.length();
       cl_int err;
@@ -374,87 +353,6 @@ PackedFunc OpenCLModuleNode::GetFunction(const std::string& name,
 Module OpenCLModuleCreate(std::string data, std::string fmt,
                           std::unordered_map<std::string, FunctionInfo> fmap, std::string source) {
   auto n = make_object<OpenCLModuleNode>(data, fmt, fmap, source);
-  n->Init();
-  return Module(n);
-}
-
-void OpenCLSPIRVModuleNode::SaveToFile(const std::string& file_name, const std::string& format) {
-  // TODO(masahi): How SPIRV binaries should be save to a file?
-  LOG(FATAL) << "Not implemented.";
-}
-
-void OpenCLSPIRVModuleNode::SaveToBinary(dmlc::Stream* stream) {
-  stream->Write(fmap_);
-  stream->Write(shaders_);
-}
-
-void OpenCLSPIRVModuleNode::Init() {
-  workspace_ = GetGlobalWorkspace();
-  workspace_->Init();
-  // initialize the kernel id, need to lock global table.
-  std::lock_guard<std::mutex> lock(workspace_->mu);
-  for (const auto& kv : fmap_) {
-    const std::string& key = kv.first;
-    KTRefEntry e;
-    if (workspace_->free_kernel_ids.size() != 0) {
-      e.kernel_id = workspace_->free_kernel_ids.back();
-      workspace_->free_kernel_ids.pop_back();
-    } else {
-      e.kernel_id = workspace_->num_registered_kernels++;
-    }
-    e.version = workspace_->timestamp++;
-    kid_map_[key] = e;
-  }
-
-  // zero initialize cl_program pointers for each device kernel
-  for (auto& kv : shaders_) {
-    programs_.insert({kv.first, std::vector<cl_program>(workspace_->devices.size(), nullptr)});
-  }
-}
-
-cl_kernel OpenCLSPIRVModuleNode::InstallKernel(cl::OpenCLWorkspace* w, cl::OpenCLThreadEntry* t,
-                                               const std::string& func_name, const KTRefEntry& e) {
-  std::lock_guard<std::mutex> lock(build_lock_);
-  int device_id = t->device.device_id;
-  if (programs_[func_name][device_id] == nullptr) {
-    auto it = shaders_.find(func_name);
-    const unsigned char* s = (const unsigned char*)it->second.data.data();
-    size_t len = it->second.data.size() * sizeof(uint32_t);
-    cl_int err;
-    cl_device_id dev = w->devices[device_id];
-    auto platform = w->device_to_platform[dev];
-    programs_[func_name][device_id] =
-        clCreateProgramWithBinary(w->contexts[platform], 1, &dev, &len, &s, nullptr, &err);
-    OPENCL_CHECK_ERROR(err);
-
-    // build program
-    err = clBuildProgram(programs_[func_name][device_id], 1, &dev, nullptr, nullptr, nullptr);
-
-    if (err != CL_SUCCESS) {
-      size_t len;
-      std::string log;
-      clGetProgramBuildInfo(programs_[func_name][device_id], dev, CL_PROGRAM_BUILD_LOG, 0, nullptr,
-                            &len);
-      log.resize(len);
-      clGetProgramBuildInfo(programs_[func_name][device_id], dev, CL_PROGRAM_BUILD_LOG, len,
-                            &log[0], nullptr);
-      LOG(FATAL) << "OpenCL build error for device=" << dev << "\n" << log;
-    }
-  }
-  // build kernel
-  cl_int err;
-  cl_kernel kernel = clCreateKernel(programs_[func_name][device_id], func_name.c_str(), &err);
-  OPENCL_CHECK_ERROR(err);
-  t->kernel_table[e.kernel_id].kernel = kernel;
-  t->kernel_table[e.kernel_id].version = e.version;
-  kernels_.push_back(kernel);
-  return kernel;
-}
-
-Module OpenCLModuleCreate(const std::unordered_map<std::string, SPIRVShader>& shaders,
-                          const std::string& spirv_text,
-                          std::unordered_map<std::string, FunctionInfo> fmap) {
-  auto n = make_object<OpenCLSPIRVModuleNode>(shaders, spirv_text, fmap);
   n->Init();
   return Module(n);
 }
