+++ src/runtime/vulkan/vulkan_wrapped_func.h
@@ -29,8 +29,8 @@
 
 #include "../meta_data.h"
 #include "../pack_args.h"
-#include "../thread_storage_scope.h"
 #include "../spirv/spirv_shader.h"
+#include "../thread_storage_scope.h"
 #include "vulkan/vulkan_core.h"
 #include "vulkan_common.h"
 #include "vulkan_device.h"
